Run Babel at build time instead of runtime to produce standalone ES5 scripts.
Run Babel as part of the build instead of a runtime require hook. Remove the
runtime dependency on babel-core so the scripts can run without requiring
the babel-register runtime require hook, as is described in the Babel docs
as 'not suitable for libraries'.

Move ES6 source files to src directories to separate them from the Babel
outputs, and point package.json files to the Babel output scripts under lib.

Adjust the build tools to the src and lib locations and the removal of the
runtime dependency on babel-core.

Add the necessary .npmrc and .gitignore files to all the packages.