Cleanup usage and reporting data models
Rename metrics to measured_usage in submitted usage to reflect the fact
that submitted usage contains raw measurements and not metrics, which are
the result of metering these raw usage measurements.

Rename unit to measure in measured_usage and metric in metered_usage, as
the term unit was incorrectly used to name measures and metrics (a unit is
the type of a measure or a metric, not its name.)

Change measure and metric names to lowercase as they're property names and
not constants.

Fix the metric lookup code to lookup metrics by name instead of relying
on their production order.

Clean up the storage resource definition example, which now only uses
Javascript metering functions instead of a mix of Javascript and the old (
now deprecated) metering formula DSL. The analytics resource definition
example is now only using the metering formula DSL as these formulas are
still supported, although deprecated.

Change the reporting service to use the same data model as the other services
instead of its own GraphQL model.

See trackers [#101098030] and [#101700906].