Pass usage start and end times to meter functions
Meter functions that compute time-based consumption metrics (eg. cpu or
memory used by apps and containers and metered on a time basis) need the
usage measurement start and end times to calculate the usage metric.

This change adds the start and end times from the usage document to the
measurement object passed to the meter functions.

See tracker [#101018516].