Add time-based metering to tests and sample resources
Add a sample 'linux-container' resource that uses a memory consumption
metric (similar to the metrics that app runtime metering may need to use)
to illustrate how to implement a form of time-based consumption metering.

Add a test-resource that combines both discrete event metering (e.g. api
calls) and time-based consumption metering (memory) and adjust metering
and aggregation unit tests to use that test-resource.

Haven't adjusted the integration tests yet as they're not running in the
build and I'm not able to run them individually to test them. Will open
an issue to report that.

Change the accumulator service to keep track of previous and current
accumulated values instead of deltas. The aggregate functions are now
passed the previous and current accumulated usage instead of a usage
delta, allowing service providers to better control how to aggregate
accumulating usage.

Simplify some of the sample and test metering and aggregation functions,
and change the samples to show that they can be written as actual functions
(converted later to strings) instead of strings.

Add support for an 'istanbul ignore file' directive to the mocha runner to
disable Istanbul instrumentation in resource config files.

Contributes to the following tracker user stories:
- [#101097784] (rating of time-based metrics),
- [#101097398] (consistent APIs for discrete and time-based metrics)
- [#101700906] (consistent reporting of discrete and time-based metrics)
- [#101018516] (potential basis for time-based app metering)