Separate the db partitioning between state and log dbs
The functionality of the logdbs have not changed.
Statedbs on the otherhand have separated from logdbs.
The epoch of a statedb is just a single epoch. Since
the statedb is storing a document forever for that
particular instance/organization, there is no need
to have epochs based upon the month/day. Only
accumulator has its statedb changed to this format
as it is the only one currently that stores all
7 time dimensions.

See Pivotal Tracker [#101098368]