Change accumulator document ids to use raw end time
Instead of using the beginning/end of the day for the
document key in the accumulator, the id has changed
to use the unaltered time. To be able to query the
usage on a smaller time dimension than a day, the
end time needs to be unchanged. Otherwise, all
hours/minutes/seconds would all fall under the same
query due to being in the same day.

See Pivotal Tracker [#101098368]