Add Multiple Time Window Logic to Accumulator
Quantity has been changed into a 7-length array of objects.
The object itself has not changed. Each element in the array
stand for a different time dimension in when the usage occured.
The format is [second, minute, hour, day, month, year, forever].
The usage will be accumulated to each time dimension in
the array. This will allow querying/reporting in more than just
the day.

To calculate if usage moves to a new time window, the end time
is converted into a format of YYYYMMDDHHmmSS. When comparing
the accumulated usage's end time to the new usage's end time,
both are converted into the above format and then divided
by their respective scale(ie: 100 for minutes or 10000 for
hours), and if the difference is greater than one, the usage
has moved onto a new window in that particular time dimension.

The aggregator has been adjusted to work with the forever
element from the accumulated usage, until support for
different time windows has been added to it

See Pivotal Tracker [#101098368]