Implement resource config API
Abacus now uses a REST API to retrieve resource configs. This will allow
integrators of Abacus to implement their own management of resource
configs. They just need to provide a provisioning service endpoint
implementing that API and point Abacus to it.

An example implementation of the resource config API is provided in the
Abacus provisioning stub.

All Abacus services have been adjusted to use the new resource config API.
This required a bit of refactoring of the services as the API is (like all
networked APIs, asynchronous.)

See tracker [#101097454].