Validate resource_ids and usage time in provisioning service
Move validation of resource_ids to the provisioning service now that
it's responsible for managing resources configured by resource
providers and providing a resource config API to the rest of the
Abacus services.

Add usage time to the provisioning validation API to allow the service
implementer to match the usage time with the resource config effective
time.

See tracker [#101097454].