Implement resource pricing config API
Abacus now uses a REST API to retrieve resource pricing config. This
will allow integrators of Abacus to implement their own management of
resource pricing data. They just need to provide an account service
endpoint implementing that API and point Abacus to it.

An example implementation of the resource pricing config API is provided
in the Abacus account stub service.

The rating service has been adjusted to use the new resource pricing
config API. This required a bit of refactoring of that service as the API
is (like all networked APIs, asynchronous.)

See tracker [#101097454].