Multiple Time Window Support For Rate & Report
Following the pattern for accumulator and aggregator, rate
and report now have support for the 7-length array that
was used in quantity. Rate slightly changes the structure
of the metric inside of aggregated usage. Replacing the
quantity and cost properties with a windows property that
is a 7-length array containing an object with cost and
quantity.

Report also adds onto that logic by providing charge
and summary where applicable. Charge on the higher levels
have been replaced with a windows object as well.

Report's query logic has changed to do a descending query
on the provided timestamp up to the previous year instead
of looking for the beginning of the day time of
the given timestamp. The graphql schema has been adjusted
for the new structure.

Report integration tests have been disabled until they've
been adjusted to the new format.

See Pivotal Tracker [#101098368]