Fix quantity logic in Accumulator Integration Tests
Instead of using a naive catch-all function that works
off of conditions of u, the quantity is calculated
via a function that is passed in which allows the
different metrics to be accurate on how their
quantity is calculated. This also solves the issue
with multiple usage records not working for storage.