Support opening a DB without automatically creating it.
Pass PouchDB options down to the DB constructor function used by
our dbclient module. This allows the caller to specify the skip_setup
option for example to avoid creating a DB just by opening it, useful
in particular if you're only going to do reads on that DB.

Add the given DB read/write operation mode to the key used to cache
DB handles.

Part of the fix for issue #69.