Fix the summary/charge functions in linux-container
The summary and charge functions were dividing the cost/quantity
by an hour in milliseconds. However, the functions were not
dividing the consumed/burned values by the same value, so
there were two different timescales being added together. The
division has been moved to the outside of the rest of the
function so the values are properly scaled to the hour.