Implement handling of delayed usage submissions
User can configure a slack window with environment
variable SLACK. The format is /^[0-9][MDhms]$/. This
will keep a history of usage up to configured point
and only to the precision of the configured scale.
The first element in a window array is the current
quantity/cost and the following elements are each
previous window. This will allow viewing history
of a particular window and late usage submissions.

See Pivotal Tracker [#105330208]