Improve implementation of duplicate detection.
Adjust the keys of the ouput docs in some of the services to make
duplicate detection work correctly.

Add a dedupe module that provides efficient duplicate detection without
having to check the db except for about 1 out of 100 false positives.

Adjust some of the test cases which were incorrectly generating
duplicate usage docs.

Partial fix for #100.