Skip checks outside of the current window in bridge itests
Since the cfbridge itests are checking the current
window, sometimes the test could pass over the
time window in which it was submitted for. Reports
and submissions which fall into different time windows
should be fine if charge/summary/cost/quantity is 0.
This should resolve occasional failures as described
in issue #104