Make it easier to config nb of db partitions
The number of db partitions used by a service and the number
of apps used to deploy the next service in the Abacus pipeline
can now be configured separately using two different environment
variables.

Use DB_PARTITIONS for the number of db partitions to be used by
the current service.
Use SINK_PARTITIONS for the number of apps used to deploy the next
(sink) service.

Fixes #114