Simplify addressing of multiple app instances
Use a a more organized list of port numbers with ranges that won't
overlap when apps get configured to run with multiple instances.

Document the list of default test ports in a section of doc/conf.md.

Support configuration of the lower bound of a port range for apps
running multiple instances through an env variable instead of having
to hardcode it in the apps.

See #132