Lower resource and pricing LRU cache timeout
No need to cache them for a whole day as we're checking
their effective dates in increments of 10 min anyway.

The pricing country doesn't need to be cached for a whole
day either, a few hours is sufficient to minimize the calls
to the account service, and a with a lower timeout the cache
will consume less memory.