Throttle calls to REST posts and database operations
See #139. Throttling to a lower number of concurrent calls helps
limit the number of unnecessary Node context switches and also
prevents overloading the target service and database with too
many concurrent calls. Will need to make the default number
configurable with an env variable later. For now 100 concurrent
calls seems a reasonable default given that the services can
be deployed as multiple concurrent apps and/or instances to
scale out if needed.