Make charge/summary functions window-aware
Charge and summary functions now take in a from and to parameter
that is the beginning and ending of the particular window the
function is running on. This will prevent summary/charge from
bleeding their calculations into the next windows in time-based
metrics. This is part of the fix for issue #182