Add provider, resource, plan ids to collected usage keys
This will help in all the cases where a user or usage processing app
needs to retrieve usage collected from a specific provider, resource
and plan, for example to rerate, fixup, or in general resubmit that
usage.

This will also make it easy to find all the runtime usage submitted
on a given month and resubmit it the next month to address issue #88.