A pipeline config that skips most intermediate Abacus DBs
The smalldb config configures the Abacus pipeline to skip most of
the usage DBs and only store usage in the collected, accumulated
and aggregated DBs.

That config can be useful for environments where a full audit of
all the usage processing steps is not needed for example. You can
then reduce the amount of DB space used by Abacus by not storing all
the inputs and outputs of all the intermediate usage processing steps.

To try it do this:
npm start smalldb
npm run demo