Optional downsampling of accumulated and aggregated usage history
Add a SAMPLING config option that allows the accumulated and aggregated
usage histories to be stored in the db in hourly or daily docs instead
of a new document per submitted usage.

Useful to save db space if needed in space constrained deployment.
See issue #345 for more background on that down sampling approach.

Fixes #345.