Support doc updates on MongoDB
Adjust the mongoclient module to provide support for updates, _rev and
rev fields compatible with the couchclient module.

Put now uses updateOne if the given doc contains a _rev field.

Batch_put and bulkDocs now use a MongoDB Bulk op with either updateOnes
for updates and insertOnes for inserts.

Documents returned by geti, batch_get, and allDocs now return a _rev field
set to '1' to allow the caller to detect that the doc comes from a version
of the doc in the DB (as opposed to a new doc that's not been written to
the DB yet).

Put, batch_put, remove, batch_remove and bulkDocs now return a rev field
set to '1' in their results, to allow the caller to use that rev field
and propagate it to updates of the retrieved doc.

Fixes #347.