Remove duplicate copy of resource instance usage
Use correct doc seqid in references to resource instance accumulated
usage docs.

Dataflow module now passes the id of the original input doc to the
service mapper/reducer function. That id is now also present in the
doc written to the input db, as that'll help keep track of the
chain of docs representing an end to end processing flow.

Fixes #337.