Support zero delay batch config with BATCH_DELAY=0
This will help the team experiment with processing call batches without
waiting for a batch to fill up in situations with very sporadic calls (no
need for batching then) or high loads (no need to wait to really fill up
the batches as most batches will already be reasonably filled with in
flight requests).

Also, when the batch delay is configured to 0 we now use setImmediate
instead of setTimeout, to help see how that'll affect the scheduling of
batches configured that way.