Fail when creating or updating a service failed (#611)
In createOrUpdateService error checks where completely turned off. However this should only be the case for the check of the existance of the service via the "cf service" command. Otherwise errors during "cf create-service" or "cf update-service" go unnoticed. 
Even when these commands trigger an asynchronous action the return code is 0, which allows error checks to be enabled during the execution of these commands.